# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AqarMatch API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ListingDealType = typeof ListingDealType[keyof typeof ListingDealType];


export const ListingDealType = {
  بيع: 'بيع',
  إيجار: 'إيجار',
} as const;

export interface Listing {
  id: number;
  deal_type: ListingDealType;
  wilaya: string;
  municipality: string;
  neighborhoods: string[];
  asking_price: number;
  user_phone: string;
  created_at: string;
  expires_at: string;
  is_active: boolean;
  days_remaining: number;
}

export type ListingInputDealType = typeof ListingInputDealType[keyof typeof ListingInputDealType];


export const ListingInputDealType = {
  بيع: 'بيع',
  إيجار: 'إيجار',
} as const;

export interface ListingInput {
  deal_type: ListingInputDealType;
  wilaya: string;
  municipality: string;
  neighborhoods: string[];
  asking_price: number;
  floor_price: number;
  user_phone: string;
}

export interface RenewRequest {
  seller_phone: string;
}

export interface MatchRequest {
  budget: number;
  buyer_phone: string;
}

export interface MatchResult {
  matched: boolean;
  listing_id: number;
  /** @nullable */
  seller_phone?: string | null;
  message?: string;
}

export interface WilayaCount {
  wilaya: string;
  count: number;
}

export interface ListingStats {
  total: number;
  for_sale: number;
  for_rent: number;
  by_wilaya: WilayaCount[];
  recent_listings: Listing[];
}

export interface Wilaya {
  code: number;
  name: string;
  name_ar: string;
}

export interface MatchAttempt {
  id: number;
  buyer_phone: string;
  budget: number;
  matched: boolean;
  created_at: string;
}

export interface ListingWithInquiries {
  id: number;
  deal_type: string;
  wilaya: string;
  municipality: string;
  neighborhoods: string[];
  asking_price: number;
  created_at: string;
  expires_at: string;
  is_active: boolean;
  days_remaining: number;
  matches: MatchAttempt[];
  total_inquiries: number;
  matched_count: number;
}

export interface SellerInquirySummary {
  phone: string;
  total_listings: number;
  total_inquiries: number;
  total_matched: number;
  listings: ListingWithInquiries[];
}

export type GetListingsParams = {
deal_type?: string;
wilaya?: string;
municipality?: string;
max_price?: number;
};